persona_support_agent/requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
streamlit
tiktoken
